requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
djangorestframework==3.6.3
gcoin==1.1.47
idna==2.5
pysha3==1.0.2
pytz==2017.2
requests==2.17.3
rlp==0.5.1
urllib3==1.21.1
